api/requirements.txt
backcall==0.1.0
click==7.1.2
decorator==4.4.2
Flask==1.1.2
Flask-HTTPAuth==4.0.0
flask-marshmallow==0.12.0
Flask-SQLAlchemy==2.4.1
ipdb==0.13.2
ipython==7.14.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.17.0
Jinja2==2.11.2
MarkupSafe==1.1.1
marshmallow==3.5.2
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.5
psycopg2-binary==2.8.5
ptyprocess==0.6.0
Pygments==2.6.1
PyJWT==1.4.2
six==1.14.0
SQLAlchemy==1.3.16
traitlets==4.3.3
wcwidth==0.1.9
Werkzeug==1.0.1
